Improve handling of incoming updates with context
The changes include the introduction of context to functions in the files tdlib.go, manager.go, and README.md. This is done for better management of incoming update processes. The docker file 'Dockerfile-Alpine' was also modified to properly install the 'tzdata' package. The introduction of the context helps to handle system signals and shutdown procedures and improves error propagation when processes are terminated.